+++ OpenSim/Moco/MocoUtilities.h
@@ -145,74 +145,6 @@ OSIMMOCO_API MocoTrajectory simulateTrajectoryWithTimeStepping(
         const MocoTrajectory& trajectory, Model model,
         double integratorAccuracy = SimTK::NaN);
 
-/// The map provides the index of each state variable in
-/// SimTK::State::getY() from its each state variable path string.
-/// Empty slots in Y (e.g., for quaternions) are ignored.
-/// @ingroup mocoutil
-OSIMMOCO_API
-std::vector<std::string> createStateVariableNamesInSystemOrder(
-        const Model& model);
-
-#ifndef SWIG
-/// Same as above, but you can obtain a map from the returned state variable
-/// names to the index in SimTK::State::getY() that accounts for empty slots
-/// in Y.
-/// @ingroup mocoutil
-OSIMMOCO_API
-std::vector<std::string> createStateVariableNamesInSystemOrder(
-        const Model& model, std::unordered_map<int, int>& yIndexMap);
-
-/// The map provides the index of each state variable in
-/// SimTK::State::getY() from its state variable path string.
-/// @ingroup mocoutil
-OSIMMOCO_API
-std::unordered_map<std::string, int> createSystemYIndexMap(const Model& model);
-#endif
-
-/// Create a vector of control names based on the actuators in the model for
-/// which appliesForce == True. For actuators with one control (e.g.
-/// ScalarActuator) the control name is simply the actuator name. For actuators
-/// with multiple controls, each control name is the actuator name appended by
-/// the control index (e.g. "/actuator_0"); modelControlIndices has length equal
-/// to the number of controls associated with actuators that apply a force
-/// (appliesForce == True). Its elements are the indices of the controls in the
-/// Model::updControls() that are associated with actuators that apply a force.
-/// @ingroup mocoutil
-OSIMMOCO_API
-std::vector<std::string> createControlNamesFromModel(
-        const Model& model, std::vector<int>& modelControlIndices);
-/// Same as above, but when there is no mapping to the modelControlIndices.
-/// @ingroup mocoutil
-OSIMMOCO_API
-std::vector<std::string> createControlNamesFromModel(const Model& model);
-/// The map provides the index of each control variable in the SimTK::Vector
-/// returned by Model::getControls(), using the control name as the
-/// key.
-/// @throws Exception if the order of actuators in the model does not match
-///     the order of controls in Model::getControls(). This is an internal
-///     error, but you may be able to avoid the error by ensuring all Actuator%s
-///     are in the Model's ForceSet.
-/// @ingroup mocoutil
-OSIMMOCO_API
-std::unordered_map<std::string, int> createSystemControlIndexMap(
-        const Model& model);
-
-/// Throws an exception if the order of the controls in the model is not the
-/// same as the order of the actuators in the model.
-/// @ingroup mocoutil
-OSIMMOCO_API void checkOrderSystemControls(const Model& model);
-
-/// Throws an exception if the same label appears twice in the list of labels.
-/// The argument copies the provided labels since we need to sort them to check
-/// for redundancies.
-/// @ingroup mocoutil
-OSIMMOCO_API void checkRedundantLabels(std::vector<std::string> labels);
-
-/// Throws an exception if any label in the provided list does not match any
-/// state variable names in the model.
-OSIMMOCO_API void checkLabelsMatchModelStates(const Model& model,
-        const std::vector<std::string>& labels);
-
 /// Convert a trajectory covering half the period of a symmetric motion into a
 /// trajectory over the full period. This is useful for simulations of half a
 /// gait cycle.
